Refactor `create_patient` Into Shared Clinical Module
This change refactors the `create_patient` function into the new shared
clinical module. Logic flow was changed slightly to keep existing functionality
and more easily allow for extensibility with other data ingestion in the future.
